Cambio tipo de datos de operaciones a enteros

+++ calculadoraTP1/src/operations.c
@@ -3,7 +3,7 @@
 #include "operations.h"
 
 
-int sumNumbers(float number1, float number2, float *result){
+int sumNumbers(int number1, int number2, int *result){
 	if(number1 && number2){
 		//Averiguar que es desreferenciar una variable
 		*result = number1 + number2;
@@ -12,7 +12,7 @@ int sumNumbers(float number1, float number2, float *result){
 	return -1;
 }
 
-int restNumbers(float number1, float number2, float *result){
+int restNumbers(int number1, int number2, int *result){
 	if(number1 && number2){
 		*result = number1 - number2;
 		return 0;
@@ -20,7 +20,7 @@ int restNumbers(float number1, float number2, float *result){
 	return -1;
 }
 
-int divideNumbers(float number1, float number2, float *result){
+int divideNumbers(int number1, int number2, float *result){
 	if(number1 && number2 && number2 != 0){
 		*result = number1 / number2;
 		return 0;
@@ -28,7 +28,7 @@ int divideNumbers(float number1, float number2, float *result){
 	return -1;
 }
 
-int multiplyNumbers(float number1, float number2, float *result){
+int multiplyNumbers(int number1, int number2, int *result){
 	if(number1 && number2){
 		*result = number1 * number2;
 		return 0;
@@ -36,7 +36,7 @@ int multiplyNumbers(float number1, float number2, float *result){
 	return -1;
 }
 
-int numberFactorial(float number, float *result){
+int numberFactorial(int number, int *result){
 	if(number){
 		int i;
 		float factorial;
